improve validator manager kind retrieval
